[!7][DEV] Prepare for PyPi release
# Why is the change needed?

To release the package on PyPi, the tags in the classifiers should be chosen among [those available on PyPi](https://pypi.org/classifiers/) otherwise the release fails. In addition, the README should include the mention to the possibility of installing from pip.

# What changes does the patch introduce?

Fixes the classifiers, extends the README, and automatically get the versioning in the code.

# How was this patch tested?

manual runs